remove shared dispatchers and rename to serve module
