+++ src/idom/__init__.py
@@ -1,7 +1,6 @@
 from . import config, html, log, types, web
 from .core import hooks
 from .core.component import component
-from .core.dispatcher import Stop
 from .core.events import event
 from .core.hooks import (
     create_context,
@@ -14,6 +13,7 @@ from .core.hooks import (
     use_state,
 )
 from .core.layout import Layout
+from .core.serve import Stop
 from .core.vdom import vdom
 from .sample import run_sample_app
 from .server.any import run
