+++ tests/test_server/test_common/test_multiview.py
@@ -1,33 +1,21 @@
 import pytest
 
 import idom
+from idom.server import all_implementations
 from idom.server import fastapi as idom_fastapi
 from idom.server import flask as idom_flask
 from idom.server import sanic as idom_sanic
 from idom.server import starlette as idom_starlette
-from idom.server import tornado as idom_tornado
-from idom.server.prefab import multiview_server
 from idom.testing import ServerMountPoint
 from tests.driver_utils import no_such_element
 
 
 @pytest.fixture(
-    params=[
-        # add new PerClientStateServer implementations here to
-        # run a suite of tests which check basic functionality
-        idom_fastapi.PerClientStateServer,
-        idom_flask.PerClientStateServer,
-        idom_sanic.PerClientStateServer,
-        idom_starlette.PerClientStateServer,
-        idom_tornado.PerClientStateServer,
-    ],
+    params=list(all_implementations()),
     ids=lambda cls: f"{cls.__module__}.{cls.__name__}",
 )
-def server_mount_point(request):
-    with ServerMountPoint(
-        request.param,
-        mount_and_server_constructor=multiview_server,
-    ) as mount_point:
+async def server_mount_point(request):
+    async with ServerMountPoint(request.param) as mount_point:
         yield mount_point
